% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{API_access}
\alias{API_access}
\title{Gain API Access}
\usage{
API_access(
  x = "b2026849c62827d3c5313156d93d22cef8062791",
  y = "3084b91d33dcdbdc3f566b2f214af5a6"
)
}
\arguments{
\item{x}{private key}

\item{y}{public key}
}
\value{
list of encoded API
}
\description{
This function asks users to input the API Key for Marvel.com
This function is the foundation of the entire package, without an API, the functions will not be able to gain access to Marvel information. Users need to use this function to input API key in order to call out other functions in the package.
}
\examples{
API_access("b2026849c62827d3c5313156d93d22cef8062791","3084b91d33dcdbdc3f566b2f214af5a6")

}
