% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{basic_info_by_ID}
\alias{basic_info_by_ID}
\title{Character Basic Marvel number of release information}
\usage{
basic_info_by_ID(x)
}
\arguments{
\item{x}{Character ID if you already know the ID number. If not, select from the saved out put of get_all_nameid()}
}
\value{
dataframe of the basic Marvel number of release information of the character
}
\description{
This function allows user to access the most basic information of a character. Ex: number of comics/series/stories/events avaialbale on Marvel.
}
\examples{
params <- API_access()
allnames <- get_all_nameid()
basic_info_by_ID(allnames$id[1])


}
