% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{get_all_nameid}
\alias{get_all_nameid}
\title{Get All Marvel Character Name/ID}
\usage{
get_all_nameid()
}
\value{
a dataframe consisting 3 columns : Name, ID and URL
}
\description{
This function is used to fetch all the current Marvel Characters including: Name, Character ID, and website URL
}
\details{
This function does not require any parameters as it loops through Marvel website to GET all the characters and the corresponding ID. Moreover, it will check if this character has a wiki website that features the specific information of the character.
If it does, it will return the website link, and if it doesn't, it will return string "Not Found"
}
\examples{
params <- API_access()
get_all_nameid()

}
