% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{get_character_bio}
\alias{get_character_bio}
\title{Get Character Biography information}
\usage{
get_character_bio(i, o)
}
\arguments{
\item{i}{the saved dataframe of the output get_all_nameid()}

\item{o}{the index of the character within the get_all_nameid() dataframe}
}
\value{
a dataframe of the character information
}
\description{
This function retrieves the basic information such as height, family etc of a character.
}
\details{
This function can only be used after get_all_nameid()
}
\examples{
params <- API_access()
allnames <- get_all_nameid()
get_character_bio(allnames,123)


}
