% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{get_event}
\alias{get_event}
\title{Get Marvel recent events}
\usage{
get_event(x)
}
\arguments{
\item{x}{number of events you wish to retrieve, range from 0 to 100}
}
\value{
a data frame of rencent (x) number of events
}
\description{
This function provides information of the recent Marvel events.
}
\examples{
params <- API_access()
get_event(30)

}
