% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{marvel_char_compare}
\alias{marvel_char_compare}
\title{Compare Marvel Character}
\usage{
marvel_char_compare(x, y, z)
}
\arguments{
\item{x}{first character name}

\item{y}{second character name}

\item{z}{third character name (optional)}
}
\value{
dataframe
}
\description{
This function allows user to compare the previous output of basic_info_by_ID() with different characters on Marvel. It will take at most three arguments, with the third being optional.
}
\examples{
params <- API_access()
allnames <- get_all_nameid()
marvel_char_compare(allnames$name[1],allnames$name[2],allnames$name[3])

}
