% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{search_character_comic}
\alias{search_character_comic}
\title{Search Character's comics}
\usage{
search_character_comic(x, y)
}
\arguments{
\item{x}{the saved dataframe of the output get_all_nameid()}

\item{y}{the index of the character within the get_all_nameid() dataframe}
}
\value{
dataframe of the character's comics
}
\description{
This function allows user to access the detailed comics information of a particular character
}
\examples{
params <- API_access()
allnames <- get_all_nameid()
search_character_comic(allnames,123)

}
