% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_nameid.R
\name{search_character_events}
\alias{search_character_events}
\title{Search Character's events}
\usage{
search_character_events(x, y)
}
\arguments{
\item{x}{the saved dataframe of the output get_all_nameid()}

\item{y}{the index of the character within the get_all_nameid() dataframe}
}
\value{
dataframe of the character's events
}
\description{
This function allows user to access the detailed events information of a particular character
}
\examples{
params <- API_access()
allnames <- get_all_nameid()
search_character_events(allnames,123)
}
